fix: при нажатии на крестик во время загрузки вакансий всё равно происходило выполнение запроса fix: при стирании текста в поле ввода поиск всё равно запускался
