+++ app/src/main/java/ru/practicum/android/diploma/utils/Debounce.kt
@@ -1,0 +1,26 @@
+package ru.practicum.android.diploma.utils
+
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
+
+fun <T> debounce(
+    delayMillis: Long,
+    coroutineScope: CoroutineScope,
+    useLastParam: Boolean,
+    action: (T) -> Unit
+): (T) -> Unit {
+    var debounceJob: Job? = null
+    return { param: T ->
+        if (useLastParam) {
+            debounceJob?.cancel()
+        }
+        if (debounceJob?.isCompleted != false || useLastParam) {
+            debounceJob = coroutineScope.launch {
+                delay(delayMillis)
+                action(param)
+            }
+        }
+    }
+}
